requirements.txt
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.0.0
cryptography==43.0.3
filelock==3.18.0
idna==3.10
packaging==24.2
platformdirs==4.3.7
protobuf==5.29.4
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.30.0
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.13.0
tzlocal==5.3.1
urllib3==1.26.20
